move prepare sankey data into sankey component
